Add more realistic blueprint example. (fixup)
