+++ examples/blueprints2-example/blueprint_editor_utilities.cpp
@@ -1,26 +1,27 @@
-# include "crude_blueprint_utilities.h"
+# include "blueprint_editor_utilities.h"
 # define IMGUI_DEFINE_MATH_OPERATORS
 # include <imgui_internal.h>
+# include "imgui_extras.h"
 # include <algorithm>
 # include <time.h>
 
-ImEx::IconType crude_blueprint_utilities::PinTypeToIconType(PinType pinType)
+blueprint_editor_utilities::IconType blueprint_editor_utilities::PinTypeToIconType(PinType pinType)
 {
     switch (pinType)
     {
-        case PinType::Void:     return ImEx::IconType::Circle;
-        case PinType::Any:      return ImEx::IconType::Diamond;
-        case PinType::Flow:     return ImEx::IconType::Flow;
-        case PinType::Bool:     return ImEx::IconType::Circle;
-        case PinType::Int32:    return ImEx::IconType::Circle;
-        case PinType::Float:    return ImEx::IconType::Circle;
-        case PinType::String:   return ImEx::IconType::Circle;
+        case PinType::Void:     return IconType::Circle;
+        case PinType::Any:      return IconType::Diamond;
+        case PinType::Flow:     return IconType::Flow;
+        case PinType::Bool:     return IconType::Circle;
+        case PinType::Int32:    return IconType::Circle;
+        case PinType::Float:    return IconType::Circle;
+        case PinType::String:   return IconType::Circle;
     }
 
-    return ImEx::IconType::Circle;
+    return IconType::Circle;
 }
 
-ImVec4 crude_blueprint_utilities::PinTypeToColor(PinType pinType)
+ImVec4 blueprint_editor_utilities::PinTypeToColor(PinType pinType)
 {
     switch (pinType)
     {
@@ -36,7 +37,7 @@ ImVec4 crude_blueprint_utilities::PinTypeToColor(PinType pinType)
     return ImVec4(1.0f, 1.0f, 1.0f, 1.0f);
 }
 
-bool crude_blueprint_utilities::DrawPinValue(const PinValue& value)
+bool blueprint_editor_utilities::DrawPinValue(const PinValue& value)
 {
     switch (value.GetType())
     {
@@ -66,7 +67,7 @@ bool crude_blueprint_utilities::DrawPinValue(const PinValue& value)
     return false;
 }
 
-bool crude_blueprint_utilities::EditPinValue(Pin& pin)
+bool blueprint_editor_utilities::EditPinValue(Pin& pin)
 {
     ImEx::ScopedItemWidth scopedItemWidth{120};
 
@@ -130,7 +131,7 @@ bool crude_blueprint_utilities::EditPinValue(Pin& pin)
     return true;
 }
 
-void crude_blueprint_utilities::DrawPinValueWithEditor(Pin& pin)
+void blueprint_editor_utilities::DrawPinValueWithEditor(Pin& pin)
 {
     auto storage = ImGui::GetStateStorage();
     auto activePinId = storage->GetInt(ImGui::GetID("PinValueEditor_ActivePinId"), false);
@@ -174,7 +175,7 @@ void crude_blueprint_utilities::DrawPinValueWithEditor(Pin& pin)
 
 
 
-const crude_blueprint_utilities::vector<crude_blueprint_utilities::Node*> crude_blueprint_utilities::GetSelectedNodes(Blueprint& blueprint)
+const blueprint_editor_utilities::vector<blueprint_editor_utilities::Node*> blueprint_editor_utilities::GetSelectedNodes(Blueprint& blueprint)
 {
     auto selectedObjects = ax::NodeEditor::GetSelectedNodes(nullptr, 0);
 
@@ -194,7 +195,7 @@ const crude_blueprint_utilities::vector<crude_blueprint_utilities::Node*> crude_
     return result;
 }
 
-const crude_blueprint_utilities::vector<crude_blueprint_utilities::Pin*> crude_blueprint_utilities::GetSelectedLinks(Blueprint& blueprint)
+const blueprint_editor_utilities::vector<blueprint_editor_utilities::Pin*> blueprint_editor_utilities::GetSelectedLinks(Blueprint& blueprint)
 {
     auto selectedObjects = ax::NodeEditor::GetSelectedLinks(nullptr, 0);
 
@@ -214,7 +215,7 @@ const crude_blueprint_utilities::vector<crude_blueprint_utilities::Pin*> crude_b
     return result;
 }
 
-crude_blueprint_utilities::PinValueBackgroundRenderer::PinValueBackgroundRenderer(const ImVec4 color, float alpha /*= 0.25f*/)
+blueprint_editor_utilities::PinValueBackgroundRenderer::PinValueBackgroundRenderer(const ImVec4 color, float alpha /*= 0.25f*/)
 {
     m_DrawList = ImGui::GetWindowDrawList();
     m_Splitter.Split(m_DrawList, 2);
@@ -223,17 +224,17 @@ crude_blueprint_utilities::PinValueBackgroundRenderer::PinValueBackgroundRendere
     m_Alpha = alpha;
 }
 
-crude_blueprint_utilities::PinValueBackgroundRenderer::PinValueBackgroundRenderer()
+blueprint_editor_utilities::PinValueBackgroundRenderer::PinValueBackgroundRenderer()
     : PinValueBackgroundRenderer(ImGui::GetStyleColorVec4(ImGuiCol_CheckMark))
 {
 }
 
-crude_blueprint_utilities::PinValueBackgroundRenderer::~PinValueBackgroundRenderer()
+blueprint_editor_utilities::PinValueBackgroundRenderer::~PinValueBackgroundRenderer()
 {
     Commit();
 }
 
-void crude_blueprint_utilities::PinValueBackgroundRenderer::Commit()
+void blueprint_editor_utilities::PinValueBackgroundRenderer::Commit()
 {
     if (m_Splitter._Current == 0)
         return;
@@ -259,24 +260,24 @@ void crude_blueprint_utilities::PinValueBackgroundRenderer::Commit()
     m_Splitter.Merge(m_DrawList);
 }
 
-void crude_blueprint_utilities::PinValueBackgroundRenderer::Discard()
+void blueprint_editor_utilities::PinValueBackgroundRenderer::Discard()
 {
     if (m_Splitter._Current == 1)
         m_Splitter.Merge(m_DrawList);
 }
 
-crude_blueprint_utilities::DebugOverlay::DebugOverlay(Blueprint& blueprint)
+blueprint_editor_utilities::DebugOverlay::DebugOverlay(Blueprint& blueprint)
     : m_Blueprint(&blueprint)
 {
     m_Blueprint->SetContextMonitor(this);
 }
 
-crude_blueprint_utilities::DebugOverlay::~DebugOverlay()
+blueprint_editor_utilities::DebugOverlay::~DebugOverlay()
 {
     m_Blueprint->SetContextMonitor(nullptr);
 }
 
-void crude_blueprint_utilities::DebugOverlay::Begin()
+void blueprint_editor_utilities::DebugOverlay::Begin()
 {
     m_CurrentNode = m_Blueprint->CurrentNode();
     m_NextNode = m_Blueprint->NextNode();
@@ -290,7 +291,7 @@ void crude_blueprint_utilities::DebugOverlay::Begin()
     m_Splitter.Split(m_DrawList, 2);
 }
 
-void crude_blueprint_utilities::DebugOverlay::End()
+void blueprint_editor_utilities::DebugOverlay::End()
 {
     if (nullptr == m_CurrentNode)
         return;
@@ -298,7 +299,7 @@ void crude_blueprint_utilities::DebugOverlay::End()
     m_Splitter.Merge(m_DrawList);
 }
 
-void crude_blueprint_utilities::DebugOverlay::DrawNode(const Node& node)
+void blueprint_editor_utilities::DebugOverlay::DrawNode(const Node& node)
 {
     if (nullptr == m_CurrentNode)
         return;
@@ -326,13 +327,13 @@ void crude_blueprint_utilities::DebugOverlay::DrawNode(const Node& node)
 
     auto color = ImGui::GetStyleColorVec4(isCurrentNode ? ImGuiCol_PlotHistogram : ImGuiCol_NavHighlight);
 
-    ImEx::DrawIcon(m_DrawList, markerMin, markerMax, ImEx::IconType::FlowDown, true, ImColor(color), 0);
+    DrawIcon(m_DrawList, markerMin, markerMax, IconType::FlowDown, true, ImColor(color), 0);
 
     // Switch back to normal layer
     m_Splitter.SetCurrentChannel(m_DrawList, 0);
 }
 
-void crude_blueprint_utilities::DebugOverlay::DrawInputPin(const Pin& pin)
+void blueprint_editor_utilities::DebugOverlay::DrawInputPin(const Pin& pin)
 {
     auto flowPinValue = m_Blueprint->GetContext().GetPinValue(m_CurrentFlowPin);
     auto flowPin = flowPinValue.GetType() == PinType::Flow ? flowPinValue.As<FlowPin*>() : nullptr;
@@ -368,7 +369,7 @@ void crude_blueprint_utilities::DebugOverlay::DrawInputPin(const Pin& pin)
         if (isCurrentFlowPin)
         {
             auto iconSize = ImVec2(ImGui::GetFontSize(), ImGui::GetFontSize());
-            ImEx::Icon(iconSize, ImEx::IconType::Flow, true, color);
+            Icon(iconSize, IconType::Flow, true, color);
         }
         else
             bg.Discard();
@@ -385,7 +386,7 @@ void crude_blueprint_utilities::DebugOverlay::DrawInputPin(const Pin& pin)
     m_Splitter.SetCurrentChannel(m_DrawList, 0);
 }
 
-void crude_blueprint_utilities::DebugOverlay::DrawOutputPin(const Pin& pin)
+void blueprint_editor_utilities::DebugOverlay::DrawOutputPin(const Pin& pin)
 {
     if (nullptr == m_CurrentNode)
         return;
@@ -411,7 +412,7 @@ void crude_blueprint_utilities::DebugOverlay::DrawOutputPin(const Pin& pin)
         if (m_CurrentFlowPin.m_Id == pin.m_Id)
         {
             auto iconSize = ImVec2(ImGui::GetFontSize(), ImGui::GetFontSize());
-            ImEx::Icon(iconSize, ImEx::IconType::Flow, true, color);
+            Icon(iconSize, IconType::Flow, true, color);
         }
         else
             bg.Discard();
@@ -422,7 +423,7 @@ void crude_blueprint_utilities::DebugOverlay::DrawOutputPin(const Pin& pin)
     m_Splitter.SetCurrentChannel(m_DrawList, 0);
 }
 
-void crude_blueprint_utilities::DebugOverlay::OnEvaluatePin(const Context& context, const Pin& pin)
+void blueprint_editor_utilities::DebugOverlay::OnEvaluatePin(const Context& context, const Pin& pin)
 {
     auto nodeEditor = ax::NodeEditor::GetCurrentEditor();
     if (!nodeEditor)
@@ -435,443 +436,23 @@ void crude_blueprint_utilities::DebugOverlay::OnEvaluatePin(const Context& conte
 
 
 
-void crude_blueprint_utilities::OverlayLogger::Log(LogLevel level, const char* format, ...)
-{
-    va_list args;
-    va_start(args, format);
-    ImGuiTextBuffer textBuffer;
-    textBuffer.appendfv(format, args);
-    va_end(args);
-
-    auto timeStamp = time(nullptr);
-
-    char formattedTimeStamp[100] = {};
-    strftime(formattedTimeStamp, sizeof(formattedTimeStamp), "%H:%M:%S", localtime(&timeStamp));
-
-    char levelSymbol = ' ';
-    switch (level)
-    {
-        case LogLevel::Verbose: levelSymbol = 'V'; break;
-        case LogLevel::Info:    levelSymbol = 'I'; break;
-        case LogLevel::Warning: levelSymbol = 'W'; break;
-        case LogLevel::Error:   levelSymbol = 'E'; break;
-    }
-
-    auto formattedMessage = string("[") + formattedTimeStamp + "] [" + levelSymbol + "] " + textBuffer.c_str();
-
-    Entry entry;
-    entry.m_Level     = level;
-    entry.m_Timestamp = timeStamp;
-    entry.m_Text      = std::move(formattedMessage);
-
-    // Parse log message to get ranges for text coloring
-    entry.m_ColorRanges = ParseMessage(entry.m_Level, entry.m_Text.c_str());
-
-    m_Entries.push_back(std::move(entry));
-}
-
-void crude_blueprint_utilities::OverlayLogger::Update(float dt)
-{
-    if (m_HoldTimer)
-        return;
-
-    dt = ImMin(dt, 1.0f); // clamp dt, to prevent logs from suddenly disappearing while returning from debugger
-
-    for (auto& entry : m_Entries)
-    {
-        if (entry.m_IsPinned)
-            continue;
-
-        entry.m_Timer += dt;
-    }
-
-    auto lastIt = std::remove_if(m_Entries.begin(), m_Entries.end(), [this](const Entry& entry)
-    {
-        return entry.m_Timer >= m_MessageLifeDuration;
-    });
-
-    m_Entries.erase(lastIt, m_Entries.end());
-}
-
-void crude_blueprint_utilities::OverlayLogger::Draw(const ImVec2& a, const ImVec2& b)
-{
-    const auto canvasMin = a + ImVec2(m_Padding, m_Padding);
-    const auto canvasMax = b - ImVec2(m_Padding, m_Padding);
-    const auto canvasSize = canvasMax - canvasMin;
-
-    auto drawList = ImGui::GetWindowDrawList();
-    //drawList->AddRect(a, b, IM_COL32(255, 255, 0, 255));
-
-    m_HoldTimer = false;
-
-    if (canvasSize.x < 1.0f || canvasSize.y < 1.0f)
-        return;
-
-    const auto clipRect = ImVec4(canvasMin.x, canvasMin.y, canvasMax.x, canvasMax.y);
-
-    //drawList->AddRect(canvasMin, canvasMax, IM_COL32(255, 0, 0, 255));
-
-    float outlineSize = m_OutlineSize;
-
-    // Draw messages from the last to first
-    int entryIndex = 0;
-    auto cursor = ImVec2(canvasMin.x, canvasMax.y);
-    for (auto it = m_Entries.rbegin(), itEnd = m_Entries.rend(); it != itEnd; ++it, ++entryIndex)
-    {
-        auto& entry = *it;
-        auto message = entry.m_Text.data();
-        auto messageEnd = entry.m_Text.data() + entry.m_Text.size();
-
-        auto outlineOffset = ImCeil(outlineSize);
-
-        auto itemSize = ImGui::CalcTextSize(message, messageEnd, false, canvasSize.x);
-        itemSize.x += outlineOffset * 2.0f;
-        itemSize.y += outlineOffset * 2.0f;
-
-        cursor.y -= itemSize.y;
-
-        auto isHovered = ImGui::IsMouseHoveringRect(cursor, cursor + itemSize, true);
-
-        auto textPosition = cursor + ImVec2(outlineOffset, outlineOffset);
-
-        auto alpha = 1.0f;
-        if (entry.m_Timer > m_MessagePresentationDuration && m_MessageFadeOutDuration > 0.0f)
-        {
-            alpha = ImMin(1.0f, (entry.m_Timer - m_MessagePresentationDuration) / m_MessageFadeOutDuration);
-            alpha = 1.0f - alpha * alpha; // Make fade out non-linear
-        }
-
-        auto textColor    = m_LogTextColor;
-        auto outlineColor = m_LogOutlineColor;
-
-           textColor.Value.w = alpha;
-        outlineColor.Value.w = alpha / 2.0f;
-
-        if (isHovered || entry.m_IsPinned)
-            drawList->AddRectFilled(cursor, cursor + itemSize, isHovered ? m_HighlightFill : m_PinFill);
-
-        if (outlineSize > 0.0f)
-        {
-            // Shadow `outlineSize`. AddText() does not support subpixel rendering,
-            // to avoid artifact apply rounded up offset.
-            float outlineSize = outlineOffset;
-
-            // Poor man outline, just stack same text around place when normal text would be
-            drawList->AddText(nullptr, 0.0f, textPosition + ImVec2(+outlineSize, +outlineSize), outlineColor, message, messageEnd, canvasSize.x, &clipRect);
-            drawList->AddText(nullptr, 0.0f, textPosition + ImVec2(+outlineSize, -outlineSize), outlineColor, message, messageEnd, canvasSize.x, &clipRect);
-            drawList->AddText(nullptr, 0.0f, textPosition + ImVec2(-outlineSize, -outlineSize), outlineColor, message, messageEnd, canvasSize.x, &clipRect);
-            drawList->AddText(nullptr, 0.0f, textPosition + ImVec2(-outlineSize, +outlineSize), outlineColor, message, messageEnd, canvasSize.x, &clipRect);
-            drawList->AddText(nullptr, 0.0f, textPosition + ImVec2(+outlineSize, 0.0f), outlineColor, message, messageEnd, canvasSize.x, &clipRect);
-            drawList->AddText(nullptr, 0.0f, textPosition + ImVec2(-outlineSize, 0.0f), outlineColor, message, messageEnd, canvasSize.x, &clipRect);
-            drawList->AddText(nullptr, 0.0f, textPosition + ImVec2(0.0f, +outlineSize), outlineColor, message, messageEnd, canvasSize.x, &clipRect);
-            drawList->AddText(nullptr, 0.0f, textPosition + ImVec2(0.0f, -outlineSize), outlineColor, message, messageEnd, canvasSize.x, &clipRect);
-        }
-
-        for (int i = 0; i < 2; ++i) // Draw text twice to make it look sharper
-        {
-            auto stringStart = drawList->VtxBuffer.Size;
-            drawList->AddText(nullptr, 0.0f, textPosition, textColor, message, messageEnd, canvasSize.x, &clipRect);
-            for (auto& range : entry.m_ColorRanges)
-               TintVertices(drawList, stringStart, range.m_Color, alpha, range.m_Start, range.m_Size);
-        }
-
-        if (isHovered)
-            m_HoldTimer = true;
-
-        if (isHovered || entry.m_IsPinned)
-            drawList->AddRect(cursor, cursor + itemSize, isHovered ? m_HighlightBorder : m_PinBorder);
-
-        if (entry.m_IsPinned || isHovered)
-        {
-            ImGui::SetItemAllowOverlap();
-            ImGui::SetCursorScreenPos(cursor);
-            ImGui::PushID(entryIndex);
-            if (ImGui::InvisibleButton("##pin", itemSize))
-            {
-                entry.m_IsPinned = !entry.m_IsPinned;
-                entry.m_Timer    = 0.0f;
-            }
-            ImGui::PopID();
-
-            if (ImGui::IsItemHovered(ImGuiHoveredFlags_None))
-                ImGui::SetTooltip("Click to %s message", entry.m_IsPinned ? "unpin" : "pin");
-        }
-
-        if (cursor.y < canvasMin.y)
-            break;
-    }
-}
-
-void crude_blueprint_utilities::OverlayLogger::AddKeyword(string_view keyword)
-{
-    m_Keywords.push_back(keyword.to_string());
-}
-
-void crude_blueprint_utilities::OverlayLogger::RemoveKeyword(string_view keyword)
-{
-    auto it = std::find(m_Keywords.begin(), m_Keywords.end(), keyword);
-    if (it != m_Keywords.end())
-        m_Keywords.erase(it);
-}
-
-ImColor crude_blueprint_utilities::OverlayLogger::GetLevelColor(LogLevel level) const
-{
-    switch (level)
-    {
-        case LogLevel::Verbose: return m_LogVerboseColor; break;
-        case LogLevel::Info:    return m_LogInfoColor;    break;
-        case LogLevel::Warning: return m_LogWarningColor; break;
-        case LogLevel::Error:   return m_LogErrorColor;   break;
-    }
-
-    return ImColor();
-}
-
-void crude_blueprint_utilities::OverlayLogger::TintVertices(ImDrawList* drawList, int firstVertexIndex, ImColor color, float alpha, int rangeStart, int rangeSize)
-{
-    // We tint data vertices emitted by ImDrawList::AddText().
-    //
-    // We work under assumption that each character is built from
-    // four vertices.
-
-    const int c_VertexPerCharacter = 4;
-
-    color.Value.w = alpha;
-    auto packedColor = static_cast<ImU32>(color);
-
-    auto vertexData    = drawList->VtxBuffer.Data;
-    auto vertexDataEnd = vertexData + drawList->VtxBuffer.Size;
-
-    auto data    = vertexData + firstVertexIndex + rangeStart * c_VertexPerCharacter;
-    auto dataEnd = data + rangeSize * c_VertexPerCharacter;
-
-    if (data >= vertexDataEnd)
-        return;
-
-    if (dataEnd >= vertexDataEnd)
-        dataEnd = vertexDataEnd;
-
-    while (data < dataEnd)
-    {
-        data->col = packedColor;
-        ++data;
-    }
-}
-
-crude_blueprint::vector<crude_blueprint_utilities::OverlayLogger::Range> crude_blueprint_utilities::OverlayLogger::ParseMessage(LogLevel level, string_view message) const
-{
-    // Totally not optimized message parsing. It identifies numbers,
-    // strings, tags, keywords and symbols.
-
-    vector<Range> result;
-
-    auto levelColor = GetLevelColor(level);
-
-    // Color timestamp '[00:00:00] [x] ', this is present in all messages.
-    result.push_back({  0, 10, m_LogSymbolColor });
-    result.push_back({  1,  2, m_LogTimeColor });
-    result.push_back({  4,  2, m_LogTimeColor });
-    result.push_back({  7,  2, m_LogTimeColor });
-    result.push_back({ 11,  1, m_LogSymbolColor });
-    result.push_back({ 12,  1, levelColor });
-    result.push_back({ 13,  1, m_LogSymbolColor });
-
-    // Color rest of the message to level color.
-    result.push_back({ 15, static_cast<int>(message.size()) - 15, levelColor });
-
-    // More usefull variant of find_first_of().
-    // If end of the string is reached and last character is
-    // in searched set we return string size not npos.
-    auto find_first_of = [message](const char* s, size_t offset = 0) -> size_t
-    {
-        auto result = message.find_first_of(s, offset);
-        if (result != string_view::npos)
-            return result;
-
-        if (offset == message.size())
-            return message.size();
-
-        if (offset > message.size())
-            return string_view::npos;
-
-        if (!strchr(s, message.back()))
-            return string_view::npos;
-
-        return message.size();
-    };
-
-    // More usefull variant of find_first_not_of().
-    // Behavior is modified in same way as for find_first_of().
-    auto find_first_not_of = [message](const char* s, size_t offset = 0) -> size_t
-    {
-        auto result = message.find_first_not_of(s, offset);
-        if (result != string_view::npos)
-            return result;
-
-        if (offset == message.size())
-            return message.size();
-
-        if (offset > message.size())
-            return string_view::npos;
-
-        if (strchr(s, message.back()))
-            return string_view::npos;
-
-        return message.size();
-    };
-
-    auto isWhitespace = [](char c) { return c == ' ' || c == '\t' || c == '\n' || c == '\r'; };
-
-    auto symbolsWithWhitespace = " \t\n\r[]():;!*+-=<>{}\"'";
-    auto symbols               = symbolsWithWhitespace + 4;
-
-    // Start search after timestamp, there is no need to parse it over and over again.
-    const auto c_HeaderSize = 15;
-
-    // Seek for symbols
-    size_t searchStart = c_HeaderSize;
-    while (true)
-    {
-        auto index = find_first_of(symbols, searchStart);
-        if (index == string_view::npos)
-            break;
-
-        result.push_back({ static_cast<int>(index), 1, m_LogSymbolColor });
-
-        searchStart = index + 1;
-    }
-
-    // Seek for tags, text in brackets
-    searchStart = c_HeaderSize;
-    while (true)
-    {
-        auto tagStart = find_first_of("[", searchStart);
-        if (tagStart == string_view::npos)
-            break;
-
-        auto tagEnd = find_first_of("]", tagStart + 1);
-        if (tagEnd == string_view::npos)
-            break;
-
-        result.push_back({ static_cast<int>(tagStart) + 1, static_cast<int>(tagEnd - tagStart) - 1, m_LogTagColor });
-
-        searchStart = tagEnd + 2;
-    }
-
-    // Seek for tokens, matches standalone words separated by symbols
-    searchStart = c_HeaderSize;
-    while (true)
-    {
-        auto tokenStart = find_first_not_of(symbolsWithWhitespace, searchStart);
-        if (tokenStart == string_view::npos)
-            break;
-
-        auto tokenEnd = find_first_of(symbolsWithWhitespace, tokenStart + 1);
-        if (tokenEnd == string_view::npos)
-            tokenEnd = message.size();
-
-        auto token = message.substr(tokenStart, tokenEnd - tokenStart);
-
-        if (std::find(m_Keywords.begin(), m_Keywords.end(), token) != m_Keywords.end())
-        {
-            result.push_back({ static_cast<int>(tokenStart), static_cast<int>(tokenEnd - tokenStart), m_LogKeywordColor });
-        }
-
-        searchStart = tokenEnd + 1;
-    }
-
-    // Seek for strings, matches "strings" and 'strings'
-    searchStart = c_HeaderSize;
-    while (true)
-    {
-        auto stringStart = find_first_of("\"'", searchStart);
-        if (stringStart >= message.size())
-            break;
-
-        char search[2] = " ";
-        search[0] = message[stringStart];
-
-        auto stringEnd = find_first_of(search, stringStart + 1);
-        if (stringEnd == string_view::npos)
-            break;
-
-        result.push_back({ static_cast<int>(stringStart), static_cast<int>(stringEnd - stringStart) + 1, m_LogStringColor });
-
-        searchStart = stringEnd + 2;
-    }
-
-    // Seek for decimal numbers
-    searchStart = c_HeaderSize;
-    while (true)
-    {
-        auto numberStart = find_first_of("0123456789.", searchStart);
-        if (numberStart == string_view::npos)
-            break;
-
-        auto numberEnd = find_first_not_of("0123456789.", numberStart + 1);
-        if (numberEnd == string_view::npos)
-            break;
-
-        if (numberStart > 0 && !isWhitespace(message[numberStart - 1]) && !strchr(symbols, message[numberStart - 1]))
-        {
-            searchStart = numberEnd + 1;
-            continue;
-        }
-
-        auto dotCount = std::count_if(message.begin() + numberStart, message.begin() + numberEnd, [](auto c) { return c == '.'; });
-        if (dotCount > 1 || (dotCount == 1 && message[numberEnd - 1] == '.'))
-        {
-            searchStart = numberEnd + 1;
-            continue;
-        }
-
-        result.push_back({ static_cast<int>(numberStart), static_cast<int>(numberEnd - numberStart), m_LogNumberColor });
-
-        searchStart = numberEnd + 1;
-    }
-
-    // Now... ImDrawList::AddText() do not render whitespaces,
-    // our ranges are build on original string. To properly
-    // match generated vertex date we need to shift and shrink
-    // ranges.
-    for (auto& range : result)
-    {
-        auto spacesBefore = std::count_if(message.begin(),                 message.begin() + range.m_Start,                isWhitespace);
-        auto spacesInside = std::count_if(message.begin() + range.m_Start, message.begin() + range.m_Start + range.m_Size, isWhitespace);
-        range.m_Start -= static_cast<int>(spacesBefore);
-        range.m_Size  -= static_cast<int>(spacesInside);
-    }
-
-    // Drop all empty ranges.
-    result.erase(std::remove_if(result.begin(), result.end(), [](const auto& range)
-    {
-        return range.m_Size <= 0;
-    }), result.end());
-
-    return result;
-}
-
-
-
-
-
-crude_blueprint_utilities::ItemBuilder::ItemBuilder()
+blueprint_editor_utilities::ItemBuilder::ItemBuilder()
     : m_InCreate(ax::NodeEditor::BeginCreate(ImGui::GetStyleColorVec4(ImGuiCol_NavHighlight)))
 {
 
 }
 
-crude_blueprint_utilities::ItemBuilder::~ItemBuilder()
+blueprint_editor_utilities::ItemBuilder::~ItemBuilder()
 {
     ax::NodeEditor::EndCreate();
 }
 
-crude_blueprint_utilities::ItemBuilder::operator bool() const
+blueprint_editor_utilities::ItemBuilder::operator bool() const
 {
     return m_InCreate;
 }
 
-crude_blueprint_utilities::ItemBuilder::NodeBuilder* crude_blueprint_utilities::ItemBuilder::QueryNewNode()
+blueprint_editor_utilities::ItemBuilder::NodeBuilder* blueprint_editor_utilities::ItemBuilder::QueryNewNode()
 {
     if (m_InCreate && ax::NodeEditor::QueryNewNode(&m_NodeBuilder.m_PinId))
         return &m_NodeBuilder;
@@ -879,7 +460,7 @@ crude_blueprint_utilities::ItemBuilder::NodeBuilder* crude_blueprint_utilities::
         return nullptr;
 }
 
-crude_blueprint_utilities::ItemBuilder::LinkBuilder* crude_blueprint_utilities::ItemBuilder::QueryNewLink()
+blueprint_editor_utilities::ItemBuilder::LinkBuilder* blueprint_editor_utilities::ItemBuilder::QueryNewLink()
 {
     if (m_InCreate && ax::NodeEditor::QueryNewLink(&m_LinkBuilder.m_StartPinId, &m_LinkBuilder.m_EndPinId))
         return &m_LinkBuilder;
@@ -887,22 +468,22 @@ crude_blueprint_utilities::ItemBuilder::LinkBuilder* crude_blueprint_utilities::
         return nullptr;
 }
 
-bool crude_blueprint_utilities::ItemBuilder::LinkBuilder::Accept()
+bool blueprint_editor_utilities::ItemBuilder::LinkBuilder::Accept()
 {
     return ax::NodeEditor::AcceptNewItem(ImVec4(0.34f, 1.0f, 0.34f, 1.0f), 3.0f);
 }
 
-void crude_blueprint_utilities::ItemBuilder::LinkBuilder::Reject()
+void blueprint_editor_utilities::ItemBuilder::LinkBuilder::Reject()
 {
     ax::NodeEditor::RejectNewItem(ImVec4(1.0f, 0.0f, 0.0f, 1.0f));
 }
 
-bool crude_blueprint_utilities::ItemBuilder::NodeBuilder::Accept()
+bool blueprint_editor_utilities::ItemBuilder::NodeBuilder::Accept()
 {
     return ax::NodeEditor::AcceptNewItem(ImVec4(0.34f, 1.0f, 0.34f, 1.0f), 3.0f);
 }
 
-void crude_blueprint_utilities::ItemBuilder::NodeBuilder::Reject()
+void blueprint_editor_utilities::ItemBuilder::NodeBuilder::Reject()
 {
     ax::NodeEditor::RejectNewItem(ImVec4(1.0f, 0.0f, 0.0f, 1.0f));
 }
@@ -910,23 +491,23 @@ void crude_blueprint_utilities::ItemBuilder::NodeBuilder::Reject()
 
 
 
-crude_blueprint_utilities::ItemDeleter::ItemDeleter()
+blueprint_editor_utilities::ItemDeleter::ItemDeleter()
     : m_InDelete(ax::NodeEditor::BeginDelete())
 {
 
 }
 
-crude_blueprint_utilities::ItemDeleter::~ItemDeleter()
+blueprint_editor_utilities::ItemDeleter::~ItemDeleter()
 {
     ax::NodeEditor::EndDelete();
 }
 
-crude_blueprint_utilities::ItemDeleter::operator bool() const
+blueprint_editor_utilities::ItemDeleter::operator bool() const
 {
     return m_InDelete;
 }
 
-crude_blueprint_utilities::ItemDeleter::NodeDeleter* crude_blueprint_utilities::ItemDeleter::QueryDeletedNode()
+blueprint_editor_utilities::ItemDeleter::NodeDeleter* blueprint_editor_utilities::ItemDeleter::QueryDeletedNode()
 {
     if (m_InDelete && ax::NodeEditor::QueryDeletedNode(&m_NodeDeleter.m_NodeId))
         return &m_NodeDeleter;
@@ -934,7 +515,7 @@ crude_blueprint_utilities::ItemDeleter::NodeDeleter* crude_blueprint_utilities::
         return nullptr;
 }
 
-crude_blueprint_utilities::ItemDeleter::LinkDeleter* crude_blueprint_utilities::ItemDeleter::QueryDeleteLink()
+blueprint_editor_utilities::ItemDeleter::LinkDeleter* blueprint_editor_utilities::ItemDeleter::QueryDeleteLink()
 {
     if (m_InDelete && ax::NodeEditor::QueryDeletedLink(&m_LinkDeleter.m_LinkId, &m_LinkDeleter.m_StartPinId, &m_LinkDeleter.m_EndPinId))
         return &m_LinkDeleter;
@@ -942,22 +523,22 @@ crude_blueprint_utilities::ItemDeleter::LinkDeleter* crude_blueprint_utilities::
         return nullptr;
 }
 
-bool crude_blueprint_utilities::ItemDeleter::LinkDeleter::Accept()
+bool blueprint_editor_utilities::ItemDeleter::LinkDeleter::Accept()
 {
     return ax::NodeEditor::AcceptDeletedItem();
 }
 
-void crude_blueprint_utilities::ItemDeleter::LinkDeleter::Reject()
+void blueprint_editor_utilities::ItemDeleter::LinkDeleter::Reject()
 {
     ax::NodeEditor::RejectDeletedItem();
 }
 
-bool crude_blueprint_utilities::ItemDeleter::NodeDeleter::Accept(bool deleteLinks /*= true*/)
+bool blueprint_editor_utilities::ItemDeleter::NodeDeleter::Accept(bool deleteLinks /*= true*/)
 {
     return ax::NodeEditor::AcceptDeletedItem(deleteLinks);
 }
 
-void crude_blueprint_utilities::ItemDeleter::NodeDeleter::Reject()
+void blueprint_editor_utilities::ItemDeleter::NodeDeleter::Reject()
 {
     ax::NodeEditor::RejectDeletedItem();
 }
@@ -965,7 +546,7 @@ void crude_blueprint_utilities::ItemDeleter::NodeDeleter::Reject()
 
 
 
-void crude_blueprint_utilities::CreateNodeDialog::Open(Pin* fromPin)
+void blueprint_editor_utilities::CreateNodeDialog::Open(Pin* fromPin)
 {
     auto storage = ImGui::GetStateStorage();
     storage->SetVoidPtr(ImGui::GetID("##create_node_pin"), fromPin);
@@ -974,7 +555,7 @@ void crude_blueprint_utilities::CreateNodeDialog::Open(Pin* fromPin)
     m_SortedNodes.clear();
 }
 
-bool crude_blueprint_utilities::CreateNodeDialog::Show(Blueprint& blueprint)
+bool blueprint_editor_utilities::CreateNodeDialog::Show(Blueprint& blueprint)
 {
     if (!ImGui::IsPopupOpen("##create_node"))
         return false;
@@ -1032,7 +613,7 @@ bool crude_blueprint_utilities::CreateNodeDialog::Show(Blueprint& blueprint)
     return wasNodeCreated;
 }
 
-bool crude_blueprint_utilities::CreateNodeDialog::CreateLinkToFirstMatchingPin(Node& node, Pin& fromPin)
+bool blueprint_editor_utilities::CreateNodeDialog::CreateLinkToFirstMatchingPin(Node& node, Pin& fromPin)
 {
     for (auto nodePin : node.GetInputPins())
     {
@@ -1070,13 +651,13 @@ bool crude_blueprint_utilities::CreateNodeDialog::CreateLinkToFirstMatchingPin(N
 
 
 
-void crude_blueprint_utilities::NodeContextMenu::Open(Node* node /*= nullptr*/)
+void blueprint_editor_utilities::NodeContextMenu::Open(Node* node /*= nullptr*/)
 {
     ImGui::GetStateStorage()->SetVoidPtr(ImGui::GetID("##node-context-menu-node"), node);
     ImGui::OpenPopup("##node-context-menu");
 }
 
-void crude_blueprint_utilities::NodeContextMenu::Show(Blueprint& blueprint)
+void blueprint_editor_utilities::NodeContextMenu::Show(Blueprint& blueprint)
 {
     if (!ImGui::IsPopupOpen("##node-context-menu"))
         return;
@@ -1145,13 +726,13 @@ void crude_blueprint_utilities::NodeContextMenu::Show(Blueprint& blueprint)
 
 
 
-void crude_blueprint_utilities::PinContextMenu::Open(Pin* pin /*= nullptr*/)
+void blueprint_editor_utilities::PinContextMenu::Open(Pin* pin /*= nullptr*/)
 {
     ImGui::GetStateStorage()->SetVoidPtr(ImGui::GetID("##pin-context-menu-pin"), pin);
     ImGui::OpenPopup("##pin-context-menu");
 }
 
-void crude_blueprint_utilities::PinContextMenu::Show(Blueprint& blueprint)
+void blueprint_editor_utilities::PinContextMenu::Show(Blueprint& blueprint)
 {
     if (!ImGui::IsPopupOpen("##pin-context-menu"))
         return;
@@ -1173,13 +754,13 @@ void crude_blueprint_utilities::PinContextMenu::Show(Blueprint& blueprint)
 
 
 
-void crude_blueprint_utilities::LinkContextMenu::Open(Pin* pin /*= nullptr*/)
+void blueprint_editor_utilities::LinkContextMenu::Open(Pin* pin /*= nullptr*/)
 {
     ImGui::GetStateStorage()->SetVoidPtr(ImGui::GetID("##link-context-menu-pin"), pin);
     ImGui::OpenPopup("##link-context-menu");
 }
 
-void crude_blueprint_utilities::LinkContextMenu::Show(Blueprint& blueprint)
+void blueprint_editor_utilities::LinkContextMenu::Show(Blueprint& blueprint)
 {
     if (!ImGui::IsPopupOpen("##link-context-menu"))
         return;
